frontend/crop_disease_detection/requirements.txt
pip>=24.0
setuptools>=65.0.0
wheel>=0.40.0
pip>=24.0
setuptools>=65.0.0
wheel>=0.40.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch
torchvision
pillow
huggingface-hub
python-multipart
jinja2